+++ src/main/java/com/epam/esm/service/impl/TagServiceImpl.java
@@ -1,10 +1,17 @@
 package com.epam.esm.service.impl;
 
-import com.epam.esm.dao.DaoException;
 import com.epam.esm.dao.TagDao;
+import com.epam.esm.dao.exception.DaoException;
+import com.epam.esm.model.Certificate;
 import com.epam.esm.model.Tag;
-import com.epam.esm.service.CertificateServiceException;
 import com.epam.esm.service.TagService;
+import com.epam.esm.service.exception.ServiceException;
+import com.epam.esm.service.exception.certificate.CertificateAlreadyExistsException;
+import com.epam.esm.service.exception.certificate.CertificateNotFoundException;
+import com.epam.esm.service.exception.certificate.UnableToDeleteCertificateException;
+import com.epam.esm.service.exception.tag.TagAlreadyExistsException;
+import com.epam.esm.service.exception.tag.TagNotFoundException;
+import com.epam.esm.service.exception.tag.UnableToDeleteTagException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Service;
@@ -39,6 +46,9 @@ public class TagServiceImpl implements TagService {
     @Override
     public Tag add(Tag tag) throws DaoException {
         try {
+            if(tagDao.getByName(tag.getName())!= null){
+                throw new TagAlreadyExistsException("Tag already exists");
+            }
             return tagDao.add(tag);
         } catch (DataAccessException e) {
             throw new DaoException(e);
@@ -47,19 +57,31 @@ public class TagServiceImpl implements TagService {
 
     @Transactional
     @Override
-    public void delete(Tag tag, Long id) throws CertificateServiceException, DaoException {
+    public void delete(Tag tag, Long id) throws ServiceException, DaoException {
         try {
+            if(tagDao.getById(id) == null) {
+                throw new TagNotFoundException("Tag not found");
+            }
+
+            final Tag quantity = tagDao.getById(id);
+            if(quantity == null) {
+                throw new UnableToDeleteTagException("Unable to delete tag");
+            }
             tagDao.getById(id);
             tag.setId(id);
             tagDao.delete(tag);
         } catch (DaoException e) {
-            throw new CertificateServiceException(e);
+            throw new ServiceException(e);
         }
     }
 
     @Override
     public Tag getByName(String name) throws DaoException {
         try {
+            Tag tag = tagDao.getByName(name);
+            if (tag == null) {
+                throw new TagNotFoundException("Tag not found");
+            }
             return tagDao.getByName(name);
         } catch (DaoException e) {
             throw new DaoException(e);
